requirements.txt
Django==5.0.1 
djangorestframework==3.15.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.4.0
shortuuid
phonenumbers
pycountry